requirements.txt
azure-functions
azure-ai-formrecognizer